Force disable updatable apexes when override flag is set
As unfortunate as it sounds, we are unable to boot when updatable apexes are enabled in prebuilt vendor image but disabled in system.
Setting ro.apex.updatable to false in /product/build.prop when OVERRIDE_TARGET_FLATTEN_APEX is set to true works around this issue and effectively lets us boot up again.

Change-Id: I8e02c8493d01c9ed2cbdbc0180acd4b85ff4b861